+++ packages/swap/src/event-handlers/index.ts
@@ -22,6 +22,7 @@ import creditedOnChain from './swapping/creditedOnChain.js';
 import refundedOnChain from './swapping/refundedOnChain.js';
 import refundEgressIgnored from './swapping/refundEgressIgnored.js';
 import refundEgressScheduled from './swapping/refundEgressScheduled.js';
+import swapAborted from './swapping/swapAborted.js';
 import swapDepositAddressReady from './swapping/swapDepositAddressReady.js';
 import swapEgressIgnored from './swapping/swapEgressIgnored.js';
 import swapEgressScheduled from './swapping/swapEgressScheduled.js';
@@ -75,6 +76,7 @@ export const events = {
     SwapRequestCompleted: 'Swapping.SwapRequestCompleted',
     CreditedOnChain: 'Swapping.CreditedOnChain',
     RefundedOnChain: 'Swapping.RefundedOnChain',
+    SwapAborted: 'Swapping.SwapAborted',
   },
   ...genericPalletEvents('IngressEgress', [
     'BatchBroadcastRequested',
@@ -228,6 +230,10 @@ const handlers = [
         },
       ]),
     ],
+  },
+  {
+    spec: '1.11.0' as Semver,
+    handlers: [{ name: events.Swapping.SwapAborted, handler: swapAborted }],
   },
 ];